pyproject.toml
[tool.poetry]
name = "cmms"
version = "0.1.0"
description = "CMMS system for tasks management"
authors = ["Aliaksandr Dauhuchyts <a.dauhuchyts@dacpol.eu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
pytest-django = "^4.9.0"
django-cors-headers = "^4.6.0"
django-ipware = "^7.0.1"
utils = "^1.0.2"
django-widget-tweaks = "^1.5.0"
django-compat = "^1.0.15"
argon2-cffi = "^23.1.0"
psycopg2-binary = "^2.9.10"
pytest-mock = "^3.14.0"
django-select2 = "^8.2.3"
freezegun = "^1.5.1"
whitenoise = "^6.9.0"


[tool.poetry.group.dev.dependencies]
autoflake = "^2.3.1"
black = "^24.10.0"
django-stubs = "^5.1.1"
djangorestframework-stubs = "^3.15.1"
flake8-annotations-complexity = "^0.0.8"
flake8-bandit = "^4.1.1"
flake8-black = "^0.3.6"
flake8-bugbear = "^24.10.31"
flake8-builtins = "^2.5.0"
flake8-cognitive-complexity = "^0.1.0"
flake8-comprehensions = "^3.16.0"
flake8-debugger = "^4.1.2"
flake8-executable = "^2.1.3"
flake8-expression-complexity = "^0.0.11"
flake8-fixme = "^1.1.1"
flake8-multiline-containers = "^0.0.19"
flake8-mutable = "^1.2.0"
flake8-pep3101 = "^2.1.0"
flake8-print = "^5.0.0"
flake8-printf-formatting = "^1.1.2"
flake8-quotes = "^3.4.0"
flake8-simplify = "^0.21.0"
flake8-variables-names = "^0.0.6"
httpie = "^3.2.4"
ipython = "^8.30.0"
mypy = "^1.13.0"
pep8-naming = "^0.14.1"
flake8 = "6.0.0"
pyproject-flake8 = "6.0.0.post1"
types-requests = "^2.32.0.20241016"
pytest = "^8.3.4"
flake8-import-order = "^0.18.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
